Move fixtures in conftest to modules
- if a fixture is only used in one module, move it to that module
- where possible, avoid patching attributes of "fixture objects" and use
  factories to generate new objects instead
